refactor: codeEditor folder added
