+++ packages/blend/lib/components/CodeV2/CodeEditorV2/index.ts
@@ -1,0 +1,2 @@
+export { default as CodeEditorV2 } from './CodeEditorV2'
+export * from './codeEditorV2.types'
