+++ packages/blend/lib/context/useComponentToken.ts
@@ -61,6 +61,7 @@ import { ResponsiveStatCardV2Tokens } from '../components/StatCardV2/statcardV2.
 import { ResponsiveRadioV2Tokens } from '../components/SelectorV2/RadioV2/radioV2.tokens'
 import { ResponsivePopoverV2Tokens } from '../components/PopoverV2/popoverV2.token'
 import { ResponsiveBreadcrumbV2Tokens } from '../components/BreadcrumbV2/breadcrumbV2.tokens'
+import { ResponsiveCodeV2Tokens } from '../components/CodeV2/tokens'
 
 export const useComponentToken = (
     component: keyof ComponentTokenType
@@ -128,7 +129,8 @@ export const useComponentToken = (
     | ResponsiveRadioV2Tokens
     | ResponsivePopoverV2Tokens
     | ResponsiveTextInputV2Tokens
-    | ResponsiveBreadcrumbV2Tokens => {
+    | ResponsiveBreadcrumbV2Tokens
+    | ResponsiveCodeV2Tokens => {
     const { componentTokens } = useTheme()
     switch (component) {
         case 'TOOLTIP':
@@ -257,6 +259,8 @@ export const useComponentToken = (
             return componentTokens.RADIOV2
         case 'BREADCRUMBV2':
             return componentTokens.BREADCRUMBV2
+        case 'CODEEDITORV2':
+            return componentTokens.CODEEDITORV2
         default:
             throw new Error(`Unknown component token: ${component}`)
     }
